Filter and delete own projects

See merge request inowas/morpheus!167

+++ src/backend/src/morpheus/project/application/write/Project/DeleteProject.py
@@ -1,9 +1,10 @@
 import dataclasses
 
 from morpheus.common.types import DateTime, Uuid
-from morpheus.common.types.Exceptions import NotFoundException
+from morpheus.common.types.Exceptions import NotFoundException, InsufficientPermissionsException
 from morpheus.common.types.event_sourcing.EventEnvelope import EventEnvelope
 from morpheus.common.types.event_sourcing.EventMetadata import EventMetadata
+from morpheus.project.application.read.PermissionsReader import permissions_reader
 from morpheus.project.application.read.ProjectReader import project_reader
 from morpheus.project.application.write.CommandBase import CommandBase
 from morpheus.project.application.write.CommandHandlerBase import CommandHandlerBase
@@ -32,14 +33,18 @@ class DeleteProjectCommandHandler(CommandHandlerBase):
         if not project_reader.project_exists(command.project_id):
             raise NotFoundException(f'Project with id {command.project_id.to_str()} does not exist')
 
+        user_id = command.user_id
+
         # todo assert user has access to project
-        # permissions = permissions_reader.get_permissions(command.project_id)
+        permissions = permissions_reader.get_permissions(command.project_id)
+        if permissions.owner_id != user_id:
+            raise InsufficientPermissionsException('User does not have owner-permissions to delete project')
+
         # read user with groups (from user module)
         # pass permissions and user with groups to domain service that checks permission
 
         # delete project from filesystem
         # delete project from mongo db
-
         # delete all assets for project
         asset_handling_service.delete_all_assets_for_project(command.project_id)
